git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@2312 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/property.c
@@ -576,7 +576,7 @@ static MagickBooleanType Get8BIMProperty(const Image *image,const char *key)
     *name='\0';
   sub_number=1;
   if (*name == '#')
-    sub_number=StringToLong(&name[1]);
+    sub_number=(ssize_t) StringToLong(&name[1]);
   sub_number=MagickMax(sub_number,1L);
   resource=(char *) NULL;
   status=MagickFalse;
